refactor(cli): enumerate + crash-proof version/help/config/upgrade
Partial fix for #157 (PR C of the planned A/B/C split; closes out the CLI
crash-proof audit portion of the issue).

Completes the "CLI dispatch hygiene" ask from PR B's out-of-scope list.
Every recovery-adjacent subcommand now routes through a dedicated minimal
module that imports nothing from @temporalio/* / rxjs / @grpc/* /
nice-grpc, dispatched from src/cli.ts BEFORE the full ./cli/commands
surface loads. The test/cli-crash-proof-isolation.test.ts suite
enumerates these modules and enforces the guarantee in CI.

Retrospective finding (fixed in same PR)
----------------------------------------
PR #218's daemon-command-isolation.test.ts was silently skipping in CI
the entire time it's been green. The `before` hook's `fs.existsSync`
check computed the dist path via `path.resolve(__dirname, '..')` — but
at runtime the compiled test lives at dist-test/test/, so `..` resolves
to dist-test/, not the repo root. The distfile check therefore always
failed, `this.skip()` fired, and mocha reported "2 pending" — which
looks fine in a CI summary. PR #218 shipped with the isolation guarantee
actually unenforced; only scripts/verify-daemon-isolation-guard.js (PR
#219) actually ran the detector. Property held in practice, but the CI
test was a placebo.

Three mitigations in this PR:
  1. Correct path: path.resolve(__dirname, '..', '..')
  2. Fail-loud on missing dist (throws with a clear "run npm run build"
     message). Explicit SKIP_ISOLATION_TEST=1 env gate for dev-loop
     opt-out.
  3. Counter + after-hook: isolationAssertionsRun is incremented by each
     `it` body; the suite fails explicitly if the counter stays at 0.
     Catches any future silent-skip regression in the `before` path.

The retrospective is front-loaded in the CHANGELOG `### Fixed` section —
project history readers need to know which CI signals were meaningless
during that window.

Crash-proof entrypoints added in PR C
-------------------------------------
- `version` / `-v` / `--version` — inlined in src/cli.ts; reads
  package.json directly, no module load at all.
- `help` / `-h` / `--help` — new src/cli/help-text.ts (~100-line
  dedicated module; imports only ./output).
- `upgrade` — new src/cli/upgrade-command.ts. Temporal bits
  (@temporalio/client, ../connection) are dynamic-imported inside the
  existing active-session-warning try/catch. SDK load failure silently
  skips the warning — the already-designed soft-fail path — instead of
  crashing the whole command. Strictly better UX for the "upgrade out
  of a broken SDK" scenario that motivated #157.
- `config show` / `config set` — src/cli/config-command.ts's
  ../connection import is now dynamic-imported inside the
  interactive-setup's connection-test step. The pure set/show paths
  become crash-proof; interactive's prompt/file-write steps are
  crash-proof and only the connection test depends on the SDK.

Enumeration pattern (the architectural win)
-------------------------------------------
test/cli-crash-proof-isolation.test.ts now enumerates
CRASH_PROOF_MODULES = [{name, ts, dist}, …]. Each entry gets two
automatic checks:
  - Runtime: child-process spawn that `require`s the compiled dist and
    asserts require.cache contains zero Temporal-adjacent keys.
  - Static: source-file regex scan that matches TypeScript STATIC
    `import ... from '...'` syntax only, ignoring dynamic
    `await import(...)`, `require(varname)`, and comments / strings.
    This is what lets upgrade-command.ts legitimately reference
    '@temporalio/client' inside a dynamic import without a
    per-module exemption list.

Future crash-proof candidates are a one-line addition to
CRASH_PROOF_MODULES. Documented inline for future contributors.

Not crash-proof (intentional)
-----------------------------
- `preflight` — legitimately tests Temporal reachability; dynamic-imported
  in cli.ts so its static import no longer leaks into the module graph.
- `attachment-info` — queries a workflow; inherently Temporal-touching.
- All other verbs (start, status, restart, destroy, migrate, restore,
  pause, resume, broadcast, init, server, up, down, …) — same.

These still go through ./cli/commands, which is now the ONLY
Temporal-heavy module in cli.ts's lazy-load path.

Docs
----
New section in docs/troubleshooting.md: "Crash-proof subcommands
(recovery levers under broken Temporal SDK)" — tabulates every CLI
subcommand and its crash-proof status so users know which levers are
available during an SDK failure.

Deletions from src/cli/commands.ts
----------------------------------
- help() function (moved to help-text.ts)
- version() function (inlined in cli.ts)
- upgrade() function (moved to upgrade-command.ts)
- UpgradeOpts interface (moved with upgrade)

Tests
-----
- `npm run build` — clean; workflow-bundle.js regenerated
- `npm test` — 766 passing, 25 pending (Vitest-only — see summary),
  0 failing. Up from PR B's effective baseline of 756 real-passing
  (the PR A isolation test's 2 skipped cases + this PR's 6 new
  enumeration cases + 2 new cases per crash-proof module = +10 real
  assertions).
- Silent-skip guard verified: removing an `it` body or making the
  `before` hook skip would immediately trip the after-hook counter
  assertion.

Follow-ups (scope-excluded)
---------------------------
- Lint guard for `this.skip()` requiring an inline SKIP-REASON comment
  — per PO suggestion, file as separate issue.
- Cleaner TS-compiled-to-out-of-tree pattern for scripts/ (currently
  protected by .gitignore `!scripts/*.js` negation, fragile to
  `foo.generated.js` accidents).
