+++ src/cli/commands.ts
@@ -2661,281 +2661,12 @@ async function setPausedState(client: Client, ensemble: string, paused: boolean)
   }
 }
 
-export function help() {
-  console.log(`
-${out.bold('claude-tempo')} — Multi-session Claude Code coordination via Temporal
+// `help()`, `version()`, and `upgrade()` moved out of commands.ts (#157 PR C)
+// to keep them crash-proof under a broken Temporal SDK install:
+//   - help     → src/cli/help-text.ts         (dedicated minimal module)
+//   - version  → inlined in src/cli.ts         (package.json read only)
+//   - upgrade  → src/cli/upgrade-command.ts    (dynamic Temporal imports)
+// All three are routed directly from `src/cli.ts` before `./cli/commands`
+// is dynamic-imported, so they remain available as recovery levers.
 
-${out.bold('Getting started:')}
-  ${out.cyan('claude-tempo up')}                  Set up everything and launch a conductor
-
-${out.bold('Usage:')}
-  claude-tempo <command> [options]
-
-${out.bold('Commands:')}
-  ${out.cyan('up')}      [ensemble]    First-time setup: start Temporal, configure MCP, launch conductor
-  ${out.cyan('down')}    [ensemble]    Tear down everything: sessions, daemon, Temporal server, MCP config
-  ${out.cyan('server')}                Start the Temporal dev server and register search attributes
-  ${out.cyan('conduct')} [ensemble]    Start a conductor session (resumes existing, --replace to restart)
-  ${out.cyan('start')}   [ensemble]    Start a player session
-  ${out.cyan('stop')}    [ensemble]    Stop sessions (-n <name> for one, or --all)
-  ${out.cyan('status')}  [ensemble]    Show active sessions and Temporal health
-  ${out.cyan('ensemble')} <sub>       Manage saved ensemble lineups (save/list/show)
-  ${out.cyan('broadcast')} <message>   Send a message to all active players
-  ${out.cyan('restart')} <name>        Restart a session (reap + claim + context replay + spawn)
-  ${out.cyan('detach')} <name>         Gracefully reap a session's adapter (workflow survives)
-  ${out.cyan('destroy')} <name>        Terminally end a session workflow
-  ${out.cyan('migrate')} <name> --host Move a session to a different host
-  ${out.cyan('attachment-info')} <name> Inspect the V2 attachment phase + current holder
-  ${out.cyan('restore')} [name]         Restore orphaned session(s) — interactive picker, or --all / --from-host / --dry-run
-  ${out.cyan('release')} [ensemble]   Release all held players (unlock outbox, deliver messages)
-  ${out.cyan('pause')}   [ensemble]   Pause an ensemble (sessions, scheduler, maestro)
-  ${out.cyan('resume')}  [ensemble]   Resume a paused ensemble (add --release to also release held sessions)
-  ${out.cyan('agent-types')} <sub>    Manage player type definitions (list/show/init)
-  ${out.cyan('daemon')}    <sub>       Manage the worker daemon (start/stop/status/logs)
-  ${out.cyan('upgrade')}  [version]    Upgrade claude-tempo to latest (or specific version)
-  ${out.cyan('config')}                Configure Temporal connection settings
-  ${out.cyan('init')}                  Register MCP server globally (or --project for .mcp.json)
-  ${out.cyan('preflight')}             Run preflight checks only
-  ${out.cyan('help')}                  Show this help message
-
-${out.bold('Connection options (all commands):')}
-  --temporal-address <addr>    Temporal server address (default: localhost:7233)
-  --temporal-namespace <ns>    Temporal namespace (default: default)
-  --temporal-api-key <key>     Temporal API key (for Temporal Cloud)
-  --temporal-tls-cert <path>   Path to TLS client certificate
-  --temporal-tls-key <path>    Path to TLS client key
-
-${out.bold('Other options:')}
-  -n, --name <name>           Set the session window name (start/conduct/up only)
-  --agent <claude|copilot>    Agent type to spawn (default: from config; start/conduct)
-  --skip-preflight            Skip preflight checks (start/conduct only)
-  --background                Run Temporal in background (server only)
-  --project                   Use per-project .mcp.json instead of global (init only)
-  --keep-mcp                  Don't remove MCP config (down only)
-  --keep-daemon               Don't stop the worker daemon (down only)
-  -y, --yes                   Skip confirmation prompt (down only)
-  --all                       Stop all sessions (stop only)
-  --lineup <name|file>        Load ensemble lineup by name or file path (up/conduct)
-  --no-hold                   Skip startup hold (requires --lineup on up/conduct)
-  --ensemble <name>           Target a specific ensemble (stop/down)
-  -d, --dir <path>            Target directory (default: cwd)
-
-${out.bold('Config command:')}
-  ${out.dim('claude-tempo config')}              Interactive connection setup
-  ${out.dim('claude-tempo config show')}         Show resolved config
-  ${out.dim('claude-tempo config set <k> <v>')}  Set a config value
-
-  Settings are saved to ~/.claude-tempo/config.json.
-  Also reads ~/.config/temporalio/temporal.yaml as a fallback.
-
-  ${out.bold('Resolution order:')} CLI flag > env var > config file > temporal CLI config > default
-
-${out.bold('First time? Run this:')}
-  ${out.dim('cd your-project')}
-  ${out.dim('claude-tempo up')}
-
-${out.bold('Typical workflow:')}
-  ${out.dim('claude-tempo server')}               Start Temporal (once, keep running)
-  ${out.dim('claude-tempo conduct myband')}       Start a conductor
-  ${out.dim('claude-tempo start myband')}         Add player sessions
-  ${out.dim('claude-tempo start myband --agent copilot -n copilot-1')}   Add a Copilot player
-  ${out.dim('claude-tempo status myband')}        Check who's active
-
-${out.bold('Environment:')}
-  CLAUDE_TEMPO_ENSEMBLE       Default ensemble name (fallback: "default")
-  TEMPORAL_ADDRESS            Default Temporal address (fallback: localhost:7233)
-  TEMPORAL_NAMESPACE          Default Temporal namespace (fallback: "default")
-  TEMPORAL_API_KEY            Temporal API key
-  TEMPORAL_TLS_CERT_PATH      Path to TLS client certificate
-  TEMPORAL_TLS_KEY_PATH       Path to TLS client key
-  CLAUDE_TEMPO_DEFAULT_AGENT  Default agent type: claude or copilot (fallback: claude)
-`);
-}
-
-export function version() {
-  try {
-    const pkg = JSON.parse(readFileSync(join(PACKAGE_ROOT, 'package.json'), 'utf8'));
-    out.log(`claude-tempo v${pkg.version}`);
-  } catch {
-    out.log('claude-tempo (unknown version)');
-  }
-}
-
-// ── upgrade command ──────────────────────────────────────────────────────────
-
-interface UpgradeOpts extends CliOverrides {
-  version?: string; // target version (e.g. "0.20.0", "latest"); defaults to "latest"
-}
-
-export async function upgrade(opts: UpgradeOpts) {
-  const config = getConfig(opts);
-  const targetVersion = opts.version || 'latest';
-  const installSpec = targetVersion === 'latest' ? 'claude-tempo' : `claude-tempo@${targetVersion}`;
-
-  // Read current version
-  let currentVersion = 'unknown';
-  try {
-    const pkg = JSON.parse(readFileSync(join(PACKAGE_ROOT, 'package.json'), 'utf8'));
-    currentVersion = pkg.version || 'unknown';
-  } catch { /* ignore */ }
-
-  out.heading('claude-tempo upgrade');
-  out.log(`  Current: v${currentVersion}`);
-  out.log(`  Target:  ${targetVersion}`);
-  console.log();
-
-  // Check for active sessions — warn the user
-  let activeSessions = 0;
-  try {
-    const connection = await Promise.race([
-      createTemporalConnection(config),
-      new Promise<never>((_, reject) => setTimeout(() => reject(new Error('timeout')), 3000)),
-    ]);
-    const client = new Client({ connection, namespace: config.temporalNamespace });
-    const query = 'WorkflowType = "claudeSessionWorkflow" AND ExecutionStatus = "Running"';
-    for await (const _wf of client.workflow.list({ query })) {
-      activeSessions++;
-    }
-  } catch {
-    // Can't connect to Temporal — that's fine, proceed with upgrade
-  }
-
-  if (activeSessions > 0) {
-    out.warn(`${activeSessions} active session(s) detected. They will lose daemon connectivity during upgrade.`);
-    out.log(`  ${out.dim('Consider running: claude-tempo stop --all')}`);
-    console.log();
-  }
-
-  // Stop the daemon gracefully (releases .node file locks)
-  const daemonWasRunning = isDaemonRunning();
-  if (daemonWasRunning) {
-    out.log('Stopping daemon...');
-    stopDaemon();
-    // Wait for process to exit — important on Windows for file lock release
-    const deadline = Date.now() + 5000;
-    while (Date.now() < deadline && isDaemonRunning()) {
-      await new Promise(r => setTimeout(r, 200));
-    }
-    if (isDaemonRunning()) {
-      out.error('Daemon did not stop in time. Try: claude-tempo daemon stop');
-      process.exit(1);
-    }
-    out.success('Daemon stopped');
-  }
-
-  // Build the detached updater script.
-  // This is a self-contained Node.js script (no native module deps) that:
-  //   1. Waits for the CLI process to exit (by PID)
-  //   2. Runs npm install -g
-  //   3. Verifies the install
-  //   4. Restarts the daemon
-  const cliPid = process.pid;
-  const isWin = process.platform === 'win32';
-
-  const updaterScript = `
-const { execFileSync } = require('child_process');
-const fs = require('fs');
-
-const PID = ${cliPid};
-const INSTALL_SPEC = ${JSON.stringify(installSpec)};
-const TARGET = ${JSON.stringify(targetVersion)};
-const IS_WIN = ${isWin};
-const LOG_PATH = ${JSON.stringify(join(CLAUDE_TEMPO_HOME, 'upgrade.log'))};
-
-function log(msg) {
-  const line = new Date().toISOString() + ' ' + msg;
-  try { fs.appendFileSync(LOG_PATH, line + '\\n'); } catch {}
-  console.log(msg);
-}
-
-function isPidAlive(pid) {
-  try { process.kill(pid, 0); return true; }
-  catch { return false; }
-}
-
-async function main() {
-  // Wait for CLI process to exit (up to 10s)
-  log('Waiting for CLI process (pid ' + PID + ') to exit...');
-  const deadline = Date.now() + 10000;
-  while (Date.now() < deadline && isPidAlive(PID)) {
-    await new Promise(r => setTimeout(r, 300));
-  }
-  if (isPidAlive(PID)) {
-    log('WARNING: CLI process still alive after 10s, proceeding anyway');
-  }
-
-  // Run npm install -g
-  log('Installing ' + INSTALL_SPEC + '...');
-  try {
-    const npmCmd = IS_WIN ? 'npm.cmd' : 'npm';
-    execFileSync(npmCmd, ['install', '-g', INSTALL_SPEC], {
-      stdio: 'inherit',
-      timeout: 120000,
-    });
-    log('Install completed');
-  } catch (err) {
-    log('Install FAILED: ' + err.message);
-    log('Recovery: npm install -g ' + INSTALL_SPEC);
-    process.exit(1);
-  }
-
-  // Verify installation
-  try {
-    const tempoCmd = IS_WIN ? 'claude-tempo.cmd' : 'claude-tempo';
-    const ver = execFileSync(tempoCmd, ['--version'], {
-      encoding: 'utf8',
-      timeout: 10000,
-    }).trim();
-    log('Verified: ' + ver);
-  } catch (err) {
-    log('WARNING: Could not verify installation: ' + err.message);
-    log('Recovery: npm install -g claude-tempo');
-  }
-
-  // Restart the daemon
-  log('Restarting daemon...');
-  try {
-    const tempoCmd = IS_WIN ? 'claude-tempo.cmd' : 'claude-tempo';
-    execFileSync(tempoCmd, ['daemon', 'start'], {
-      stdio: 'inherit',
-      timeout: 30000,
-    });
-    log('Daemon restarted');
-  } catch (err) {
-    log('WARNING: Daemon restart failed: ' + err.message);
-    log('Run manually: claude-tempo daemon start');
-  }
-
-  log('Upgrade complete!');
-}
-
-main().catch(err => {
-  log('Upgrade failed: ' + err.message);
-  process.exit(1);
-});
-`.trim();
-
-  // Clear previous upgrade log before spawning
-  const logPath = join(CLAUDE_TEMPO_HOME, 'upgrade.log');
-  try { writeFileSync(logPath, ''); } catch { /* ignore */ }
-
-  // Spawn the updater as a detached child process
-  out.log(`Spawning upgrade process for ${out.cyan(installSpec)}...`);
-
-  const child = cpSpawn(process.execPath, ['-e', updaterScript], {
-    detached: true,
-    stdio: 'ignore',
-    env: { ...process.env },
-  });
-  child.unref();
-
-  console.log();
-  out.success('Upgrade started in background');
-  out.log(`  ${out.dim('Monitor progress: ')}`);
-  if (isWin) {
-    out.log(`    ${out.dim('powershell Get-Content -Wait "' + logPath + '"')}`);
-  } else {
-    out.log(`    ${out.dim('tail -f ' + logPath)}`);
-  }
-  out.log(`  ${out.dim('If it fails, run manually: npm install -g ' + installSpec)}`);
-}
+// See src/cli/upgrade-command.ts for the upgrade handler implementation.
